+++ .eslintrc.js
@@ -1,0 +1,26 @@
+module.exports = {
+  env: {
+    amd: true,
+    browser: true,
+    es6: true,
+    jquery: false,
+    node: true,
+  },
+
+  globals: {
+    'Blockly': true,
+  },
+
+  extends: [
+    "airbnb",
+  ],
+
+  parser: "babel-eslint",
+  parserOptions: {
+    sourceType: 'module',
+    ecmaVersion: 6,
+    ecmaFeatures: {
+      jsx: true,
+    },
+  },
+};