+++ src/BlocklyEditor.jsx
@@ -1,66 +1,103 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import PropTypes from 'prop-types';
 import Immutable from 'immutable';
 
 import BlocklyToolbox from './BlocklyToolbox';
 import BlocklyWorkspace from './BlocklyWorkspace';
 
-var BlocklyEditor = React.createClass({
-  propTypes: {
-    initialXml: React.PropTypes.string,
-    workspaceConfiguration: React.PropTypes.object,
-    wrapperDivClassName: React.PropTypes.string,
-    toolboxCategories: React.PropTypes.array,
-    toolboxBlocks: React.PropTypes.array,
-    xmlDidChange: React.PropTypes.func,
-	workspaceDidChange: React.PropTypes.func,
-    onImportXmlError: React.PropTypes.func,
-    processToolboxCategory: React.PropTypes.func
-  },
+const BlockPropType = PropTypes.shape({
+  type: PropTypes.string.isRequired,
+  shadow: PropTypes.bool,
+  fields: PropTypes.object,
+  values: PropTypes.object,
+  statements: PropTypes.object,
+  next: PropTypes.object,
+  mutation: PropTypes.shape({
+    attributes: PropTypes.object,
+    innerContent: PropTypes.string,
+  }),
+});
 
-  toolboxDidUpdate: function() {
-    var workspaceConfiguration = this.props.workspaceConfiguration || {};
-    if (this.refs.workspace && !workspaceConfiguration.readOnly) {
-      this.refs.workspace.toolboxDidUpdate(ReactDOM.findDOMNode(this.refs.toolbox));
-    }
-  },
+const categoryPropsNonRecursive = {
+  type: PropTypes.string.isRequired,
+  name: PropTypes.string,
+  custom: PropTypes.string,
+  colour: PropTypes.string,
+  blocks: PropTypes.arrayOf(BlockPropType),
+};
 
-  componentDidMount: function() {
+const CategoryPropType = PropTypes.shape({
+  ...categoryPropsNonRecursive,
+  categories: PropTypes.arrayOf(PropTypes.shape(categoryPropsNonRecursive)),
+});
+
+class BlocklyEditor extends React.Component {
+  static propTypes = {
+    initialXml: PropTypes.string,
+    workspaceConfiguration: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+    wrapperDivClassName: PropTypes.string,
+    toolboxCategories: PropTypes.arrayOf(CategoryPropType.isRequired),
+    toolboxBlocks: PropTypes.arrayOf(BlockPropType.isRequired),
+    xmlDidChange: PropTypes.func,
+    workspaceDidChange: PropTypes.func,
+    onImportXmlError: PropTypes.func,
+    processToolboxCategory: PropTypes.func,
+  };
+
+  static defaultProps = {
+    initialXml: null,
+    workspaceConfiguration: null,
+    wrapperDivClassName: null,
+    toolboxCategories: null,
+    toolboxBlocks: null,
+    xmlDidChange: null,
+    workspaceDidChange: null,
+    onImportXmlError: null,
+    processToolboxCategory: null,
+  };
+
+  componentDidMount = () => {
     this.toolboxDidUpdate();
-	
-	if(this.props.xmlDidChange) {
-	  if(typeof console !== 'undefined') {
-	    console.error('Warning: xmlDidChange is deprecated and will be removed in future versions! Please use workspaceDidChange instead.');
-	  }
-	}
-  },
 
-  xmlDidChange: function(newXml) {
+    if (this.props.xmlDidChange) {
+      if (typeof console !== 'undefined') {
+        // eslint-disable-next-line no-console
+        console.error('Warning: xmlDidChange is deprecated and will be removed in future versions! Please use workspaceDidChange instead.');
+      }
+    }
+  }
+
+  toolboxDidUpdate = () => {
+    const workspaceConfiguration = this.props.workspaceConfiguration || {};
+    if (this.workspace && !workspaceConfiguration.readOnly) {
+      this.workspace.toolboxDidUpdate(this.toolbox.getRootNode());
+    }
+  }
+
+  xmlDidChange = (newXml) => {
     if (this.props.xmlDidChange) {
       this.props.xmlDidChange(newXml);
     }
-  },
-  
-  workspaceDidChange: function(workspace) {
+  }
+
+  workspaceDidChange = (workspace) => {
     if (this.props.workspaceDidChange) {
       this.props.workspaceDidChange(workspace);
     }
-  },
+  }
 
-  importFromXml: function(xml) {
-    return this.refs.workspace.importFromXml(xml);
-  },
+  importFromXml = xml => this.workspace.importFromXml(xml)
 
-  resize: function() {
-    this.refs.workspace.resize();
-  },
+  resize = () => {
+    this.workspace.resize();
+  }
 
-  render: function() {
-    var toolboxMode;
+  render = () => {
+    let toolboxMode;
     if (this.props.toolboxCategories) {
-      toolboxMode = "CATEGORIES";
+      toolboxMode = 'CATEGORIES';
     } else if (this.props.toolboxBlocks) {
-      toolboxMode = "BLOCKS";
+      toolboxMode = 'BLOCKS';
     }
 
     return (
@@ -70,18 +107,21 @@ var BlocklyEditor = React.createClass({
           blocks={Immutable.fromJS(this.props.toolboxBlocks)}
           didUpdate={this.toolboxDidUpdate}
           processCategory={this.props.processToolboxCategory}
-          ref="toolbox" />
-        <BlocklyWorkspace ref="workspace"
+          ref={(toolbox) => { this.toolbox = toolbox; }}
+        />
+        <BlocklyWorkspace
+          ref={(workspace) => { this.workspace = workspace; }}
           initialXml={this.props.initialXml}
           onImportXmlError={this.props.onImportXmlError}
           toolboxMode={toolboxMode}
           xmlDidChange={this.xmlDidChange}
-		  workspaceDidChange={this.workspaceDidChange}
+          workspaceDidChange={this.workspaceDidChange}
           wrapperDivClassName={this.props.wrapperDivClassName}
-          workspaceConfiguration={this.props.workspaceConfiguration} />
+          workspaceConfiguration={this.props.workspaceConfiguration}
+        />
       </div>
     );
   }
-});
+}
 
 export default BlocklyEditor;