+++ src/BlocklyToolbox.jsx
@@ -1,55 +1,47 @@
+/* eslint-disable react/no-array-index-key */
+
 import React from 'react';
+import PropTypes from 'prop-types';
 import { is } from 'immutable';
 import ImmutablePropTypes from 'react-immutable-proptypes';
 
 import BlocklyToolboxCategory from './BlocklyToolboxCategory';
 import BlocklyToolboxBlock from './BlocklyToolboxBlock';
 
-var BlocklyToolbox = React.createClass({
-  propTypes: {
+class BlocklyToolbox extends React.Component {
+  static propTypes = {
     categories: ImmutablePropTypes.list,
     blocks: ImmutablePropTypes.list,
-    processCategory: React.PropTypes.func,
-    didUpdate: React.PropTypes.func
-  },
+    processCategory: PropTypes.func,
+    didUpdate: PropTypes.func,
+  };
 
-  renderCategories: function(categories) {
-    return categories.map(function(category, i) {
-      if (category.get('type') === 'sep') {
-        return <sep key={"sep_" + i}></sep>;
-      } else if (category.get('type') === 'search') {
-        return <search key={"search_" + i}/>;
-      } else {
-        return <BlocklyToolboxCategory
-          name={category.get('name')}
-          custom={category.get('custom')}
-          colour={category.get('colour')}
-          key={"category_" + category.get('name') + "_" + i}
-          blocks={category.get('blocks')}
-          categories={category.get('categories')} />;
-      }
-    }.bind(this));
-  },
+  static defaultProps = {
+    categories: null,
+    blocks: null,
+    processCategory: null,
+    didUpdate: null,
+  };
 
-  shouldComponentUpdate: function(nextProps, nextState) {
-    return !(is(nextProps.categories, this.props.categories) && is(nextProps.blocks, this.props.blocks));
-  },
-
-  componentDidMount: function() {
+  componentDidMount = () => {
     this.props.didUpdate();
-  },
+  }
 
-  componentDidUpdate: function(prevProps, prevState) {
+  shouldComponentUpdate = nextProps => !(
+    is(nextProps.categories, this.props.categories) && is(nextProps.blocks, this.props.blocks)
+  )
+
+  componentDidUpdate = () => {
     this.props.didUpdate();
-  },
+  }
 
-  processCategory: function(category) {
-    var processedCategory = category;
+  getRootNode = () => this.rootNode
+
+  processCategory = (category) => {
+    let processedCategory = category;
 
     if (processedCategory.has('categories')) {
-      processedCategory = category.update('categories', function(subcategories) {
-        return subcategories.map(this.processCategory);
-      }.bind(this));
+      processedCategory = category.update('categories', subcategories => subcategories.map(this.processCategory));
     }
 
     if (this.props.processCategory) {
@@ -57,23 +49,38 @@ var BlocklyToolbox = React.createClass({
     }
 
     return processedCategory;
-  },
+  }
 
-  render: function() {
+  renderCategories = categories => categories.map((category, i) => {
+    if (category.get('type') === 'sep') {
+      return <sep key={`sep_${i}`} />;
+    } else if (category.get('type') === 'search') {
+      return <search key={`search_${i}`} />;
+    }
+    return (<BlocklyToolboxCategory
+      name={category.get('name')}
+      custom={category.get('custom')}
+      colour={category.get('colour')}
+      key={`category_${category.get('name')}_${i}`}
+      blocks={category.get('blocks')}
+      categories={category.get('categories')}
+    />);
+  })
+
+  render = () => {
     if (this.props.categories) {
       return (
-        <xml style={{display: "none"}}>
+        <xml style={{ display: 'none' }} ref={(node) => { this.rootNode = node; }}>
           {this.renderCategories(this.props.categories.map(this.processCategory))}
         </xml>
       );
-    } else {
-      return (
-        <xml style={{display: "none"}}>
-          {this.props.blocks.map(BlocklyToolboxBlock.renderBlock)}
-        </xml>
-      );
     }
+    return (
+      <xml style={{ display: 'none' }} ref={(node) => { this.rootNode = node; }}>
+        {this.props.blocks.map(BlocklyToolboxBlock.renderBlock)}
+      </xml>
+    );
   }
-});
+}
 
 export default BlocklyToolbox;