+++ src/BlocklyWorkspace.jsx
@@ -1,67 +1,95 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import PropTypes from 'prop-types';
 
-var debounce = function(func, wait) {
-	var timeout;
-	return function() {
-		var context = this, args = arguments;
-		var later = function() {
-			timeout = null;
-			func.apply(context, args);
-		};
-		clearTimeout(timeout);
-		timeout = setTimeout(later, wait);
-	};
-};
+function debounce(func, wait) {
+  let timeout;
 
-var BlocklyWorkspace = React.createClass({
-  propTypes: {
-    initialXml: React.PropTypes.string,
-    workspaceConfiguration: React.PropTypes.object,
-    wrapperDivClassName: React.PropTypes.string,
-    xmlDidChange: React.PropTypes.func,
-	workspaceDidChange: React.PropTypes.func,
-    onImportXmlError: React.PropTypes.func,
-    toolboxMode: React.PropTypes.oneOf(['CATEGORIES', 'BLOCKS'])
-  },
-
-  getInitialState: function() {
-    return {
-      workspace: null,
-      xml: this.props.initialXml
+  return function debouncedFunction(...args) {
+    const context = this;
+    const later = function later() {
+      timeout = null;
+      func.apply(context, args);
     };
-  },
+    clearTimeout(timeout);
+    timeout = setTimeout(later, wait);
+  };
+}
 
-  componentDidMount: function() {
+class BlocklyWorkspace extends React.Component {
+  static propTypes = {
+    initialXml: PropTypes.string,
+    workspaceConfiguration: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+    wrapperDivClassName: PropTypes.string,
+    xmlDidChange: PropTypes.func,
+    workspaceDidChange: PropTypes.func,
+    onImportXmlError: PropTypes.func,
+    toolboxMode: PropTypes.oneOf(['CATEGORIES', 'BLOCKS']),
+  };
+
+  static defaultProps = {
+    initialXml: null,
+    workspaceConfiguration: null,
+    wrapperDivClassName: null,
+    xmlDidChange: null,
+    workspaceDidChange: null,
+    onImportXmlError: null,
+    toolboxMode: 'BLOCKS',
+  };
+
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      workspace: null,
+      xml: this.props.initialXml,
+    };
+  }
+
+  componentDidMount = () => {
     // TODO figure out how to use setState here without breaking the toolbox when switching tabs
     this.state.workspace = Blockly.inject(
-      this.refs.editorDiv,
-      Object.assign({}, (this.props.workspaceConfiguration || {}), {
-        toolbox: ReactDOM.findDOMNode(this.refs.dummyToolbox)
-      })
+      this.editorDiv,
+      {
+        ...this.props.workspaceConfiguration,
+        toolbox: this.dummyToolbox,
+      },
     );
 
     if (this.state.xml) {
       if (this.importFromXml(this.state.xml)) {
         this.xmlDidChange();
       } else {
-        this.setState({xml: null}, this.xmlDidChange);
+        this.setState({ xml: null }, this.xmlDidChange);
       }
     }
-	
-	this.state.workspace.addChangeListener(this.workspaceDidChange);
 
-    this.state.workspace.addChangeListener(debounce(function() {
-      var newXml = Blockly.Xml.domToText(Blockly.Xml.workspaceToDom(this.state.workspace));
-      if (newXml == this.state.xml) {
+    this.state.workspace.addChangeListener(this.workspaceDidChange);
+
+    this.state.workspace.addChangeListener(debounce(() => {
+      const newXml = Blockly.Xml.domToText(Blockly.Xml.workspaceToDom(this.state.workspace));
+      if (newXml === this.state.xml) {
         return;
       }
 
-      this.setState({xml: newXml}, this.xmlDidChange);
-    }.bind(this), 200));
-  },
+      this.setState({ xml: newXml }, this.xmlDidChange);
+    }, 200));
+  }
 
-  importFromXml: function(xml) {
+  componentWillReceiveProps = (newProps) => {
+    if (this.props.initialXml !== newProps.initialXml) {
+      this.setState({ xml: newProps.initialXml });
+    }
+  }
+
+  shouldComponentUpdate = () => false
+
+  componentWillUnmount = () => {
+    if (this.state.workspace) {
+      this.state.workspace.dispose();
+    }
+  }
+
+  importFromXml = (xml) => {
     try {
       Blockly.Xml.domToWorkspace(Blockly.Xml.textToDom(xml), this.state.workspace);
       return true;
@@ -71,51 +99,35 @@ var BlocklyWorkspace = React.createClass({
       }
       return false;
     }
-  },
+  }
 
-  componentWillReceiveProps: function(newProps) {
-    if (this.props.initialXml != newProps.initialXml) {
-      this.setState({xml: newProps.initialXml});
-    }
-  },
-
-  componentWillUnmount: function() {
-    if (this.state.workspace) {
-      this.state.workspace.dispose();
-    }
-  },
-
-  shouldComponentUpdate: function() {
-    return false;
-  },
-
-  xmlDidChange: function() {
+  xmlDidChange = () => {
     if (this.props.xmlDidChange) {
       this.props.xmlDidChange(this.state.xml);
     }
-  },
-  
-  workspaceDidChange: function(event) {
+  }
+
+  workspaceDidChange = () => {
     if (this.props.workspaceDidChange) {
       this.props.workspaceDidChange(this.state.workspace);
     }
-  },
+  }
 
-  toolboxDidUpdate: function(toolboxNode) {
+  toolboxDidUpdate = (toolboxNode) => {
     if (toolboxNode && this.state.workspace) {
       this.state.workspace.updateToolbox(toolboxNode);
     }
-  },
+  }
 
-  resize: function() {
+  resize = () => {
     Blockly.svgResize(this.state.workspace);
-  },
+  }
 
-  render: function() {
+  render = () => {
     // We have to fool Blockly into setting up a toolbox with categories initially;
     // otherwise it will refuse to do so after we inject the real categories into it.
-    var dummyToolboxContent;
-    if (this.props.toolboxMode === "CATEGORIES") {
+    let dummyToolboxContent;
+    if (this.props.toolboxMode === 'CATEGORIES') {
       dummyToolboxContent = (
         <category name="Dummy toolbox" />
       );
@@ -123,13 +135,16 @@ var BlocklyWorkspace = React.createClass({
 
     return (
       <div className={this.props.wrapperDivClassName}>
-        <xml style={{display: "none"}} ref="dummyToolbox">
+        <xml style={{ display: 'none' }} ref={(dummyToolbox) => { this.dummyToolbox = dummyToolbox; }}>
           {dummyToolboxContent}
         </xml>
-        <div ref="editorDiv" className={this.props.wrapperDivClassName} />
+        <div
+          className={this.props.wrapperDivClassName}
+          ref={(editorDiv) => { this.editorDiv = editorDiv; }}
+        />
       </div>
     );
   }
-});
+}
 
 export default BlocklyWorkspace;